index page cleanup 4
